requirements/test.txt
-r common.txt

pytest
pytest-cov
tox
vcrpy
vcrpy-unittest
mock
behave
pyhamcrest
